requirements.txt
dnspython==2.2.1
pycryptodome==3.15.0
pycryptodomex==3.15.0
pymongo==4.1.1
PyYAML==6.0
strgen==1.3.1
StringGenerator==0.4.4
